+++ code/polynomial/Function.java
@@ -1,5 +1,7 @@
 package polynomial;
 
+import java.util.Vector;
+
 public abstract class Function {
 
     public abstract double calculate(double x);
@@ -81,5 +83,70 @@ public abstract class Function {
 
     public double slope(double x, double percision) {
         return (calculate(x) - calculate(x - percision)) / (percision);
+    }
+
+    /**
+     * My implementation of a getSolution method. Does not work in all cases as it might get "stuck" in an area. This
+     * solutions could be fixed by instead of taking the slope, seeing if the middle and the end point are on opposite
+     * sides of the x axis. This could then also be used in order to find multiple solutions.
+     * @deprecated does not work
+     */
+    public double getSolutionBSearch(double lowerBound, double higherBound, int steps) {
+        double guess = lowerBound;
+        for (int i = 0; i < steps; i++) {
+            guess = (lowerBound + higherBound) / 2;
+            if (calculate(guess) == 0) {
+                return guess;
+            }
+            if (slope(guess, 0.00001) <= 0) {
+                if (calculate(guess) <= 0) {
+                    higherBound = guess;
+                } else {
+                    lowerBound = guess;
+                }
+            } else {
+                if (calculate(guess) <= 0) {
+                    lowerBound = guess;
+                } else {
+                    higherBound = guess;
+                }
+            }
+        }
+        return guess;
+    }
+
+    public void getSolutions(Vector<Double> solutions, double lowerBound, double higherBound, int steps) {
+        if (steps == 0 || this.calculate(lowerBound) == 0) {
+            solutions.add(lowerBound);
+        }
+
+        double guess = (lowerBound + higherBound) / 2;
+        if (java.lang.Thread.activeCount() < 10) {
+            if (lowerBound * guess <= 0 && guess * higherBound <= 0) {
+                // We know that the lower bound and the guess are on opposite sides of the x axis and that guess and the
+                // higher bound are on opposite sides of the x axis, therefore there must be a on both sides of guess
+                // if the function is continuous.
+                Thread lowerThread = new Thread(() -> getSolutions(solutions, lowerBound, guess, steps - 1));
+                lowerThread.start();
+                Thread higherThread = new Thread(() -> getSolutions(solutions, guess, higherBound, steps - 1));
+                higherThread.start();
+
+                // Do not return until both threads are finished
+                while (lowerThread.isAlive() || higherThread.isAlive()) {
+                }
+                return;
+            }
+        }
+
+        // There is no longer any need to use threads
+        if (lowerBound * guess <= 0) {
+            // We know that the lower bound and the guess are on opposite sides of the x axis, therefore there must be a
+            // solution in between them if the function is continuous.
+            getSolutions(solutions, lowerBound, guess, steps - 1);
+        } else if (guess * higherBound <= 0) {
+            // We know that the higher bound and the guess are on opposite sides of the x axis, therefore there must be
+            // a solution in between them if the function is continuous.
+            getSolutions(solutions, guess, higherBound, steps - 1);
+        }
     }
 }
